requirements.txt
requests==2.32.5
beautifulsoup4==4.14.3
python-dotenv==1.2.1
playwright==1.49.1
tinydb==4.8.0
geopy
flask
pandas
